pyproject.toml
[project]
name = "flask-ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fakeredis>=2.33.0",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=4.1.1",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-smorest>=0.46.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "marshmallow-sqlalchemy>=1.4.2",
    "psycopg[binary]>=3.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
